Focus watcher default Emacs can be hardcoded to another program.
To test requires restarting the focus watcher, and there have
been some side effects doing that, so this commit also includes
tested code to stop (kill) the focus watcher. The file watcher
may be revealed as the culprit.

Note that the mouse is disabled in the window, it appears to
have locked up. Don't panic, hit the Menu key at most three
times to re-enter default mode from Secondary mode, the mouse
should move again. This is an acid test of keyboard bindings,
a user should be able to fix without releasing the mouse.